requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.5.0
MarkupSafe==1.1.1
numpy==1.18.1
pandas==1.0.1
Pillow==7.0.0
pkg-resources==0.0.0
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2019.3
requests==2.23.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
